Refactor - Create a new file for ASTWalker unittest
